Add an abstracted database service
updated to support the database service:

add annotations to UrlRepository and update the purge method so
records with a maximum Long value can be deleted.

add NoArgsConstructor to UrlEntity so the autowiring won't complain.

add a call to generate the short url - prior to this commit the short
url was a static value.