Add the REST controller for url shortening
this commit makes phase1 feature complete and the web service can now
be started as part of the Spring Boot app.

application.yml is configured to the "dev" profile now and will use a
non-persistent H2 in memory database. This can be changed to a "test"
configuration that runs on a different port and file based H2 by
default.